Modify signals to render an HTML button.
