+++ basics/signals/package.json
@@ -44,14 +44,17 @@
     "watch:src": "tsc -w"
   },
   "dependencies": {
-    "@jupyterlab/application": "^3.0.0-rc.15",
-    "@jupyterlab/launcher": "^3.0.0-rc.15",
-    "@jupyterlab/mainmenu": "^3.0.0-rc.15",
-    "@jupyterlab/translation": "^3.0.0-rc.15",
+    "@jupyterlab/application": "^3.0.3",
+    "@jupyterlab/apputils": "^3.0.3",
+    "@jupyterlab/launcher": "^3.0.3",
+    "@jupyterlab/mainmenu": "^3.0.3",
+    "@jupyterlab/translation": "^3.0.3",
     "@lumino/algorithm": "^1.3.3",
     "@lumino/coreutils": "^1.5.3",
     "@lumino/datagrid": "^0.3.1",
-    "@lumino/disposable": "^1.4.3"
+    "@lumino/disposable": "^1.4.3",
+    "@lumino/signaling": "^1.3.3",
+    "@lumino/widgets": "^1.16.1"
   },
   "devDependencies": {
     "@jupyterlab/builder": "^3.0.0-rc.15",
@@ -63,7 +66,6 @@
     "eslint-plugin-prettier": "^3.1.2",
     "eslint-plugin-react": "^7.18.3",
     "npm-run-all": "^4.1.5",
-    "prettier": "^1.19.0",
     "rimraf": "^3.0.2",
     "typescript": "~4.1.3"
   },
